sales-ai-assistant/requirements.txt

fastapi
uvicorn
google-api-python-client
google-auth
pytz
